fix: rename the command `rm` to `remove`
